refactor: update user files location
