+++ src/endpoints/user/user.router.ts
@@ -4,7 +4,7 @@ import type { IRequest } from 'models/i-request';
 import { UserSchema } from 'models/i-user';
 import { authMiddleware } from 'middlewares/auth.middleware';
 import { validationMiddleware } from 'middlewares/validation.middleware';
-import * as userController from 'controllers/user.controller';
+import * as userController from './user.controller';
 
 const router = Router();
